refactor: rename module
Rename upm-config to upm-config-io. This is in preparation for moving the UpmConfig type and functions which act on it into own module which should be called upm-config
